refactor(ff-sys): split swresample.rs into swresample/ sub-modules
swresample.rs (1860 lines) mixed context management, format conversion,
channel layout helpers, sample format constants, and audio FIFO wrappers
in a single flat file. Split into focused sub-modules:

- swresample/mod.rs       — re-exports + integration tests
- swresample/context.rs   — SwrContext alloc/init/free/is_initialized
- swresample/convert.rs   — convert, get_delay, estimate_output_samples
- swresample/channel_layout.rs — AVChannelLayout helpers and presets
- swresample/sample_format.rs  — AVSampleFormat constants and helpers
- swresample/audio_fifo.rs     — AVAudioFifo alloc/read/write/size/free

Public API is unchanged; lib.rs requires no edits since Rust resolves
`pub mod swresample;` to swresample/mod.rs automatically.

Closes #715
